Move event emitter under utils section as it is a utility and not a component.
