+++ src/terra-dev-site/app/utilities.2/ApplicationEventEmitter.app.mdx
