+++ tests/jest/utils/application-event-emitter/index.test.js
@@ -1,4 +1,4 @@
-import ApplicationEventEmitter from '../../../src/application-event-emitter';
+import ApplicationEventEmitter from '../../../../src/utils/application-event-emitter';
 
 describe('application-event-emitter/index', () => {
   it('should export ApplicationEventEmitter', () => {
